Restructure tests folders
