+++ src/routes/__tests__/2-Calculator.test.js
@@ -1,5 +1,5 @@
 import renderer from 'react-test-renderer';
-import Calculator from '../../routes/2-Calculator';
+import Calculator from '../2-Calculator';
 
 test('Calculator route renders correctly', () => {
   const tree = renderer
