+++ src/routes/__tests__/3-Quote.test.js
@@ -1,5 +1,5 @@
 import renderer from 'react-test-renderer';
-import Quote from '../../routes/3-Quote';
+import Quote from '../3-Quote';
 
 test('Quote route renders correctly', () => {
   const tree = renderer
